Extract dev tasks
